requirements.txt
kafka-python
pyspark==3.5.0
textblob
pymongo
streamlit
faker
plotly
pandas
numpy